finished example test
